Create connected_components.py
This script analyzes a binary image represented as a 2D list of 0s (white pixels) and 1s (black pixels). It builds an adjacency graph for black pixels and identifies all disjoint (connected) black regions using depth-first search (DFS). The output includes the number of black regions and the coordinates of the pixels in each region.